saving scores; fixed user id type

+++ conf/evolutions/default/2.sql
@@ -3,7 +3,7 @@
 # --- !Ups
 
 CREATE TABLE Score (
-  id bigint(20) NOT NULL AUTO_INCREMENT,
+  id int NOT NULL AUTO_INCREMENT,
   winner varchar(255) NOT NULL,
   loser varchar(255) NOT NULL,
   winnerScore int NOT NULL,
